+++ source/helpers/pkcs7/pkcs7-check.c
@@ -31,7 +31,7 @@ char *pdf_signature_error_description(pdf_signature_error err)
 #include "mupdf/helpers/pkcs7-openssl.h"
 #include <string.h>
 
-static void pdf_format_designated_name(pdf_pkcs7_designated_name *name, char *buf, int buflen)
+static void pdf_format_designated_name(pdf_pkcs7_designated_name *name, char *buf, size_t buflen)
 {
 	int i, n;
 	const char *part[] = {
@@ -50,10 +50,10 @@ static void pdf_format_designated_name(pdf_pkcs7_designated_name *name, char *bu
 			fz_strlcat(buf, part[i], buflen);
 }
 
-void pdf_signature_designated_name(fz_context *ctx, pdf_document *doc, pdf_obj *signature, char *buf, int buflen)
+void pdf_signature_designated_name(fz_context *ctx, pdf_document *doc, pdf_obj *signature, char *buf, size_t buflen)
 {
 	char *contents = NULL;
-	int contents_len = pdf_signature_contents(ctx, doc, signature, &contents);
+	size_t contents_len = pdf_signature_contents(ctx, doc, signature, &contents);
 	pdf_pkcs7_designated_name *name = NULL;
 	fz_try(ctx)
 	{
@@ -77,7 +77,7 @@ pdf_signature_error pdf_check_digest(fz_context *ctx, pdf_document *doc, pdf_obj
 	pdf_signature_error err;
 	fz_stream *bytes = NULL;
 	char *contents = NULL;
-	int contents_len = pdf_signature_contents(ctx, doc, signature, &contents);
+	size_t contents_len = pdf_signature_contents(ctx, doc, signature, &contents);
 	fz_var(err);
 	fz_var(bytes);
 	fz_try(ctx)
@@ -101,7 +101,7 @@ pdf_signature_error pdf_check_digest(fz_context *ctx, pdf_document *doc, pdf_obj
 pdf_signature_error pdf_check_certificate(fz_context *ctx, pdf_document *doc, pdf_obj *signature)
 {
 	char *contents = NULL;
-	int contents_len = pdf_signature_contents(ctx, doc, signature, &contents);
+	size_t contents_len = pdf_signature_contents(ctx, doc, signature, &contents);
 	pdf_signature_error result;
 	fz_try(ctx)
 		result = pkcs7_openssl_check_certificate(contents, contents_len);
@@ -112,7 +112,7 @@ pdf_signature_error pdf_check_certificate(fz_context *ctx, pdf_document *doc, pd
 	return result;
 }
 
-int pdf_check_signature(fz_context *ctx, pdf_document *doc, pdf_obj *signature, char *ebuf, int ebufsize)
+int pdf_check_signature(fz_context *ctx, pdf_document *doc, pdf_obj *signature, char *ebuf, size_t ebufsize)
 {
 	int res = 0;
 
@@ -144,7 +144,7 @@ int pdf_check_signature(fz_context *ctx, pdf_document *doc, pdf_obj *signature, 
 			case PDF_SIGNATURE_ERROR_SELF_SIGNED_IN_CHAIN:
 			case PDF_SIGNATURE_ERROR_NOT_TRUSTED:
 				{
-					int len;
+					size_t len;
 					fz_strlcat(ebuf, " (", ebufsize);
 					len = strlen(ebuf);
 					pdf_signature_designated_name(ctx, doc, signature, ebuf + len, ebufsize - len);
@@ -180,7 +180,7 @@ int pdf_supports_signatures(fz_context *ctx)
 
 #else
 
-void pdf_signature_designated_name(fz_context *ctx, pdf_document *doc, pdf_obj *signature, char *buf, int buflen)
+void pdf_signature_designated_name(fz_context *ctx, pdf_document *doc, pdf_obj *signature, char *buf, size_t buflen)
 {
 	fz_throw(ctx, FZ_ERROR_GENERIC, "No OpenSSL support.");
 }
@@ -198,7 +198,7 @@ pdf_signature_error pdf_check_certificate(fz_context *ctx, pdf_document *doc, pd
 }
 
 
-int pdf_check_signature(fz_context *ctx, pdf_document *doc, pdf_obj *signature, char *ebuf, int ebufsize)
+int pdf_check_signature(fz_context *ctx, pdf_document *doc, pdf_obj *signature, char *ebuf, size_t ebufsize)
 {
 	fz_strlcpy(ebuf, "No digital signing support in this build", ebufsize);
 	return 0;
